logTimeStamp/Jia: 1. 設定button的cursor 2. atoms置於store資料夾 3. 調整countdown頁面timer, task, button排版
