+++ src/components/Modules/Countdown/ProgressBarModal/ProgressBarModal.scss
@@ -1,4 +1,4 @@
-.progressBar {
+.progress-bar {
   &__overlay {
     position: fixed;
     top: 0;
