+++ src/components/Modules/Countdown/TaskDisplay/TaskDisplay.scss
@@ -1,0 +1,32 @@
+.task-display {
+  &__wrapper {
+    display: flex;
+    flex-flow: row nowrap;
+    flex: 1 1 0%;
+    justify-content: center;
+    gap: 3rem;
+    color: $colorWhite;
+    width: 80%;
+    align-items: center;
+    @include pad {
+      flex-flow: column;
+      gap: 1rem;
+    }
+  }
+
+  &__title {
+    display: flex;
+    align-items: center;
+    text-align: center;
+    font-size: 8rem;
+    white-space: pre-wrap;
+    overflow: hidden;
+    font-weight: bold;
+    height: 6rem;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5);
+
+    @include pad {
+      font-size: 4rem;
+    }
+  }
+}
