+++ src/components/Modules/Countdown/TaskDisplay/index.ts
@@ -1,0 +1,2 @@
+export * from './TaskDisplay';
+export { default } from './TaskDisplay';
