+++ src/components/Modules/Countdown/TimerDisplay/index.ts
@@ -1,0 +1,2 @@
+export * from './TimerDisplay';
+export { default } from './TimerDisplay';
