+++ src/components/common/ProgressBar/ProgressBar.scss
@@ -1,4 +1,4 @@
-.progressBar {
+.progress-bar {
   &__segment {
     display: flex;
     font-size: 0;
