+++ src/components/common/ProgressBar/index.ts
@@ -1,1 +1,2 @@
+export * from './ProgressBar';
 export { default } from './ProgressBar';
