+++ src/pages/Countdown.tsx
@@ -3,7 +3,7 @@ import { Layout, LayoutMain, LayoutModal } from '@/layout/Layout';
 import Countdown from '@/components/Modules/Countdown';
 import InputData from '@/components/Modules/Countdown/InputData';
 import ProgressBarModal from '@/components/Modules/Countdown/ProgressBarModal';
-import { showAnalyticsModalAtom } from '@/atoms/modalAtoms';
+import { showAnalyticsModalAtom } from '@/store/atoms/modalAtoms';
 
 const PageCountdown = () => {
   const showAnalyticsModal = useAtomValue(showAnalyticsModalAtom);
